Refactor: Added GP classes as base classes and their class implementations with the SE kernel in linear.py
